Found a workaround for require() in js code
The testcase code for both javascript and python was included using
org-mode :noweb references. That meant that every source block for each
solution included all the code for the testcase class, which added
noise to the actual challenge solution code.

I had to resort to use an absolute path in the require call, since the
relative call did not seem to be able to find the module js
file. Since I am using node inside a docker image, I know that the
current folder will be mounted as a volume on the "/app" folder of the
container when running node.

Now, the two classes for running the test cases get tangled into the
testcase.js and testcase.py files. Then, in each source code block I
reference the testcase class by including the corresponding
module (with require() for javascript and import for python)